include Trusted-Library: true in jar manifests
When running the sensor-applet using plugin2 (tested in
Safari) a security warning is displayed indicating to the
user that a mix of trusted and untrusted resources are
being used in the applet. This happens when using a
classloader getResourceAsStream method to retrieve an
arch-specific native library jar.

Adding 'Trusted-Library: true' to the jar manifests fixes
this problem.

ref:
Mixing Signed and Unsigned Code
http://download.oracle.com/javase/6/docs/technotes/guides/jweb/mixed_code.html