url option to prevent use of pack200-gzip transport
If you want to force network transport of a jar 
resource without using the smaller pack200-gzip 
payload add: '.no.pack.gz' to the end of the path 
to the jar.

For example changing an archive reference in an 
applet from this:

  org/concord/otrunk/otrunk.jar

to this:

  org/concord/otrunk/otrunk.jar.no.pack.gz

will force the network transport of the uncompressed 
jar. 

In general uncompressed jars will be three times
as large as compressed jars.

In the example above the difference in network 
transport size:

178 kB: org/concord/otrunk/otrunk.jar
555 kB: org/concord/otrunk/otrunk.jar.no.pack.gz