mw library demo: ONLY includes the index page
This commit does NOT include all the cml and mml
resources the library index page references. 

This is committed to share online a bug exposed when
an html page with an MW applet that references this 
index.cml page is reloaded:

  http://jnlp.dev.concord.org/mw-library.html

On Safari on the Mac the page content is blank.